+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/SaleOrderLineServiceImpl.java
@@ -183,6 +183,9 @@ public class SaleOrderLineServiceImpl implements SaleOrderLineService {
         saleOrderLine.setDiscountAmount((BigDecimal) discounts.get("discountAmount"));
       }
       saleOrderLine.setDiscountTypeSelect((Integer) discounts.get("discountTypeSelect"));
+    } else if (!saleOrder.getTemplate()) {
+      saleOrderLine.setDiscountAmount(BigDecimal.ZERO);
+      saleOrderLine.setDiscountTypeSelect(PriceListLineRepository.AMOUNT_TYPE_NONE);
     }
 
     return price;
@@ -191,18 +194,23 @@ public class SaleOrderLineServiceImpl implements SaleOrderLineService {
   protected void fillTaxInformation(SaleOrderLine saleOrderLine, SaleOrder saleOrder)
       throws AxelorException {
 
-    TaxLine taxLine = this.getTaxLine(saleOrder, saleOrderLine);
-    saleOrderLine.setTaxLine(taxLine);
+    if (saleOrder.getClientPartner() != null) {
+      TaxLine taxLine = this.getTaxLine(saleOrder, saleOrderLine);
+      saleOrderLine.setTaxLine(taxLine);
 
-    FiscalPosition fiscalPosition = saleOrder.getClientPartner().getFiscalPosition();
+      FiscalPosition fiscalPosition = saleOrder.getClientPartner().getFiscalPosition();
 
-    Tax tax =
-        accountManagementService.getProductTax(
-            saleOrderLine.getProduct(), saleOrder.getCompany(), fiscalPosition, false);
+      Tax tax =
+          accountManagementService.getProductTax(
+              saleOrderLine.getProduct(), saleOrder.getCompany(), fiscalPosition, false);
 
-    TaxEquiv taxEquiv = Beans.get(FiscalPositionService.class).getTaxEquiv(fiscalPosition, tax);
+      TaxEquiv taxEquiv = Beans.get(FiscalPositionService.class).getTaxEquiv(fiscalPosition, tax);
 
-    saleOrderLine.setTaxEquiv(taxEquiv);
+      saleOrderLine.setTaxEquiv(taxEquiv);
+    } else {
+      saleOrderLine.setTaxLine(null);
+      saleOrderLine.setTaxEquiv(null);
+    }
   }
 
   @Override
@@ -457,6 +465,34 @@ public class SaleOrderLineServiceImpl implements SaleOrderLineService {
     if (priceList != null) {
       PriceListLine priceListLine = this.getPriceListLine(saleOrderLine, priceList);
       discounts = priceListService.getReplacedPriceAndDiscounts(priceList, priceListLine, price);
+
+      if (saleOrder.getTemplate()) {
+        Integer manualDiscountAmountType = saleOrderLine.getDiscountTypeSelect();
+        BigDecimal manualDiscountAmount = saleOrderLine.getDiscountAmount();
+        Integer priceListDiscountAmountType = (Integer) discounts.get("discountTypeSelect");
+        BigDecimal priceListDiscountAmount = (BigDecimal) discounts.get("discountAmount");
+
+        if (!manualDiscountAmountType.equals(priceListDiscountAmountType)
+            && manualDiscountAmountType.equals(PriceListLineRepository.AMOUNT_TYPE_PERCENT)
+            && priceListDiscountAmountType.equals(PriceListLineRepository.AMOUNT_TYPE_FIXED)) {
+          priceListDiscountAmount =
+              priceListDiscountAmount
+                  .multiply(new BigDecimal(100))
+                  .divide(price, 2, RoundingMode.HALF_UP);
+        } else if (!manualDiscountAmountType.equals(priceListDiscountAmountType)
+            && manualDiscountAmountType.equals(PriceListLineRepository.AMOUNT_TYPE_FIXED)
+            && priceListDiscountAmountType.equals(PriceListLineRepository.AMOUNT_TYPE_PERCENT)) {
+          priceListDiscountAmount =
+              priceListDiscountAmount
+                  .multiply(price)
+                  .divide(new BigDecimal(100), 2, RoundingMode.HALF_UP);
+        }
+
+        if (manualDiscountAmount.compareTo(priceListDiscountAmount) > 0) {
+          discounts.put("discountAmount", manualDiscountAmount);
+          discounts.put("discountTypeSelect", manualDiscountAmountType);
+        }
+      }
     }
 
     return discounts;